# feat(docker): add user-provided hooks support to Docker API