… deploy/docker/hook_manager.py
… docs/examples/docker_hooks_examples.py
… docs/md_v2/core/docker-deployment.md
… tests/docker/test_hooks_comprehensive.py