## Summary
Adds comprehensive user-provided hooks support to the Docker API, allowing users to inject custom Python functions as strings that execute at specific points in the crawling pipeline. This enables advanced customization of crawler behavior without modifying the core codebase.

Fixes #1377

## List of files changed and why
- `deploy/docker/hook_manager.py` - New module implementing UserHookManager for safe validation, compilation, and execution of user-provided hook functions with error isolation
- `deploy/docker/api.py` - Updated handle_crawl_request and handle_stream_crawl_request to integrate hooks, maintain hook_manager instance for execution tracking
- `deploy/docker/schemas.py` - Added HookConfig and CrawlRequestWithHooks models to support hooks in API requests
- `deploy/docker/server.py` - Updated /crawl and /crawl/stream endpoints to accept hooks parameter, added /hooks/info endpoint for hook discovery
- `docs/md_v2/core/docker-deployment.md` - Added comprehensive hooks documentation section with security warnings, real-world examples, and best practices
- `docs/examples/docker_hooks_examples.py` - Created example file demonstrating all 8 hooks with practical use cases (authentication, performance optimization, content extraction)
- `tests/docker/test_hooks_client.py` - Added test suite for basic hooks functionality and error handling
- `tests/docker/test_hooks_comprehensive.py` - Added comprehensive test suite covering all hook points, timeout handling, and edge cases

## How Has This Been Tested?
- Built Docker image locally using `docker compose build`
- Ran container on port 11235 and verified all endpoints
- Executed comprehensive test suite (`docs/examples/docker_hooks_examples.py`) with 100% success rate
- Tested all 8 hook points individually and in combination
- Verified error isolation - malformed hooks don't crash the crawler
- Tested timeout protection with slow-running hooks
- Validated execution tracking and statistics reporting
- Confirmed JSON serialization of all hook results
- Tested with real URLs (httpbin.org, Wikipedia, BBC, GitHub)

## Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added/updated unit tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes