@SohamKukreti 2025-09-24
I've tested the hooks functionality and everything is working correctly:

1. **API Discovery**: `/hooks/info` endpoint - ✅ Working
2. **Basic Hook Execution**: `on_page_context_created` hook - ✅ Working  
   - Response: `'hooks': {'status': {'status': 'success', 'attached_hooks': ['on_page_context_created']}`
3. **Error Handling**: Malformed hooks - ✅ Working (properly handles errors)
   - Response: `'validation_errors': [{'hook_point': 'invalid_hook_point', 'error': 'Unknown hook point. Valid points: on_browser_created, on_page_context_created, before_goto, after_goto, on_user_agent_updated, on_execution_started, before_retrieve_html, before_return_html', 'code_preview': 'invalid code'}]`
4. **Comprehensive Testing**: `docs/examples/docker_hooks_example.py` - ✅ Working
5. **Authentication Scenario**: Hook-based auth flow - ✅ Working
   - Test: Basic auth with `before_goto` hook adding Authorization header
   - Response: `"authenticated": true, "user": "user"` ✅ Auth successful
   - Hook execution: 0.004s, 100% success rate
6. **Performance Impact Test**: BBC.com with/without hooks - ✅ Working
   - With hook: 3.01s processing, 0.90MB memory delta
   - Without hook: 2.94s processing, 1.46MB memory delta
   - Result: Hook reduces memory usage by 38% (resource blocking effective)
7. **Content Extraction**: Wikipedia metadata extraction - ✅ Working  
   - Test: `after_goto` and `before_retrieve_html` hooks on Wikipedia
   - Successfully extracted Python programming language metadata from infobox  
   - Result: Complete infobox data including paradigm, developer, release info, etc.
   - Hook execution: Both hooks completed successfully (0.028s + 0.018s)


### Assessment

The implementation is solid with:
- Proper error isolation (malformed hooks don't crash crawls)
- Comprehensive validation and security measures
- All 8 hook points functioning correctly
- Good documentation and examples

**LGTM!**